--- a/go.mod
+++ b/go.mod
@@ -1,8 +1,17 @@
 module github.com/hashicorp/go-bexpr
 
-go 1.14
+go 1.18
 
 require (
 	github.com/mitchellh/pointerstructure v1.2.1
 	github.com/stretchr/testify v1.8.2
 )
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/mitchellh/mapstructure v1.4.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
+
+replace github.com/mitchellh/pointerstructure v1.2.1 => github.com/remilapeyre/pointerstructure v0.0.0-20230906004826-80d813010704