--- a/go.mod
+++ b/go.mod
@@ -13,5 +13,3 @@ require (
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
-
-replace github.com/mitchellh/pointerstructure v1.2.1 => github.com/remilapeyre/pointerstructure v0.0.0-20230906004826-80d813010704