The syntax is inspired by Sentinel [Any, All Expressions](https://docs.hashicorp.com/sentinel/language/boolexpr#any-all-expressions) and the [For Statements](https://docs.hashicorp.com/sentinel/language/loops#for-statements).

Supporting this sort of complex conditions is useful for the Vault SAML auth method.