@remilapeyre 2023-09-07
Here's the results I got for the benchmarks on a `m6i.large` EC2 instance with `make bench BENCHTIME=5s BENCHFULL_ARG="-count=10 -timeout 60m"`:


```
goos: linux
goarch: amd64
pkg: github.com/hashicorp/go-bexpr
cpu: Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz
                                          │  main.txt   │              any.txt               │
                                          │   sec/op    │   sec/op     vs base               │
Evaluate/Flat_Struct_Alt_Types/#0-2         1.501µ ± 0%   1.500µ ± 0%       ~ (p=0.565 n=10)
Evaluate/Flat_Struct_Alt_Types/#1-2         1.501µ ± 0%   1.501µ ± 1%       ~ (p=0.612 n=10)
Evaluate/Flat_Struct_Alt_Types/#38-2        1.510µ ± 0%   1.524µ ± 0%  +0.89% (p=0.000 n=10)
Evaluate/Flat_Struct_Alt_Types/#39-2        1.510µ ± 0%   1.523µ ± 0%  +0.86% (p=0.001 n=10)
Evaluate/Flat_Struct_Alt_Types/#46-2        1.520µ ± 0%   1.526µ ± 0%  +0.39% (p=0.001 n=10)
Evaluate/Flat_Struct_Alt_Types/#47-2        1.519µ ± 0%   1.529µ ± 0%  +0.63% (p=0.000 n=10)
Evaluate/Flat_Struct_Alt_Types/#52-2        1.527µ ± 0%   1.540µ ± 0%  +0.85% (p=0.001 n=10)
Evaluate/Flat_Struct_Alt_Types/#53-2        1.524µ ± 0%   1.534µ ± 0%  +0.69% (p=0.000 n=10)
Evaluate/map[string]map[string]bool/#10-2   1.151µ ± 0%   1.140µ ± 0%  -0.91% (p=0.001 n=10)
Evaluate/map[string]map[string]bool/#15-2   1.442µ ± 0%   1.441µ ± 0%       ~ (p=0.393 n=10)
Evaluate/Nested_Structs_and_Maps/#0-2       1.260µ ± 0%   1.274µ ± 0%  +1.11% (p=0.000 n=10)
Evaluate/Nested_Structs_and_Maps/#1-2       1.839µ ± 0%   1.826µ ± 0%  -0.68% (p=0.000 n=10)
Evaluate/Nested_Structs_and_Maps/#2-2       1.847µ ± 0%   1.834µ ± 0%  -0.68% (p=0.001 n=10)
Evaluate/Nested_Structs_and_Maps/#6-2       2.098µ ± 0%   2.109µ ± 1%  +0.52% (p=0.030 n=10)
Evaluate/Nested_Structs_and_Maps/#11-2      1.854µ ± 0%   1.840µ ± 0%  -0.76% (p=0.001 n=10)
Evaluate/Nested_Structs_and_Maps/#13-2      3.927µ ± 0%   3.923µ ± 1%       ~ (p=0.470 n=10)
Evaluate/Nested_Structs_and_Maps/#17-2      2.870µ ± 0%   2.867µ ± 0%  -0.14% (p=0.033 n=10)
Evaluate/Flat_Struct/#0-2                   1.658µ ± 0%   1.653µ ± 0%  -0.33% (p=0.005 n=10)
Evaluate/Flat_Struct/#2-2                   1.661µ ± 0%   1.657µ ± 0%       ~ (p=0.078 n=10)
Evaluate/Flat_Struct/#39-2                  1.667µ ± 0%   1.668µ ± 0%       ~ (p=0.574 n=10)
Evaluate/Flat_Struct/#40-2                  1.669µ ± 0%   1.668µ ± 0%       ~ (p=0.557 n=10)
Evaluate/Flat_Struct/#47-2                  1.677µ ± 0%   1.678µ ± 0%       ~ (p=0.320 n=10)
Evaluate/Flat_Struct/#48-2                  1.677µ ± 0%   1.679µ ± 0%       ~ (p=0.153 n=10)
Evaluate/Flat_Struct/#53-2                  1.692µ ± 0%   1.698µ ± 0%  +0.35% (p=0.001 n=10)
Evaluate/Flat_Struct/#54-2                  1.687µ ± 0%   1.690µ ± 0%       ~ (p=0.180 n=10)
Evaluate/Flat_Struct/#58-2                  1.696µ ± 0%   1.699µ ± 0%       ~ (p=0.053 n=10)
Evaluate/Flat_Struct/#67-2                  2.003µ ± 0%   2.013µ ± 0%  +0.50% (p=0.000 n=10)
Evaluate/Flat_Struct/#68-2                  1.897µ ± 0%   1.908µ ± 0%  +0.61% (p=0.000 n=10)
Filter/Slice-2                              3.406µ ± 0%   3.399µ ± 0%       ~ (p=0.101 n=10)
Filter/Array-2                              3.275µ ± 0%   3.267µ ± 0%  -0.23% (p=0.001 n=10)
Filter/Map-2                                3.956µ ± 1%   3.938µ ± 0%  -0.43% (p=0.001 n=10)
geomean                                     1.835µ        1.837µ       +0.10%

pkg: github.com/hashicorp/go-bexpr/grammar
                                          │  main.txt   │              any.txt               │
                                          │   sec/op    │   sec/op     vs base               │
ExpressionParsing/Contains-2                34.29µ ± 0%   34.01µ ± 0%  -0.80% (p=0.001 n=10)
ExpressionParsing/Is_Not_Empty-2            40.50µ ± 0%   40.14µ ± 0%  -0.89% (p=0.000 n=10)
ExpressionParsing/In_And_Not_Equals-2       59.83µ ± 0%   59.11µ ± 0%  -1.20% (p=0.000 n=10)
ExpressionParsing/Many_Ors-2                116.6µ ± 0%   114.7µ ± 0%  -1.68% (p=0.000 n=10)
ExpressionParsing/Lots_of_Parens-2          22.83m ± 0%   22.25m ± 0%  -2.54% (p=0.000 n=10)
ExpressionParsing/Not_Equals-2              28.99µ ± 0%   28.81µ ± 0%  -0.63% (p=0.001 n=10)
ExpressionParsing/In-2                      55.74µ ± 0%   55.17µ ± 0%  -1.02% (p=0.000 n=10)
ExpressionParsing/Not_Matches-2             37.96µ ± 0%   37.64µ ± 0%  -0.83% (p=0.001 n=10)
ExpressionParsing/Not_Contains-2            34.22µ ± 0%   33.94µ ± 0%  -0.83% (p=0.000 n=10)
ExpressionParsing/Not_In_Or_Equals-2        60.97µ ± 0%   61.77µ ± 1%  +1.32% (p=0.002 n=10)
ExpressionParsing/Big_Selectors-2           87.84µ ± 0%   89.31µ ± 1%  +1.68% (p=0.000 n=10)
ExpressionParsing/Equals-2                  27.44µ ± 0%   27.29µ ± 0%  -0.53% (p=0.000 n=10)
ExpressionParsing/Is_Empty-2                36.85µ ± 0%   36.51µ ± 0%  -0.93% (p=0.000 n=10)
ExpressionParsing/Matches-2                 34.41µ ± 0%   34.23µ ± 0%  -0.53% (p=0.000 n=10)
ExpressionParsing/Lots_of_Ops-2             132.7µ ± 0%   130.7µ ± 0%  -1.50% (p=0.000 n=10)
ExpressionParsing/Not_In-2                  59.75µ ± 0%   59.95µ ± 1%       ~ (p=0.393 n=10)
ExpressionParsing/Not_Equals_And_Equals-2   152.1µ ± 0%   149.6µ ± 0%  -1.65% (p=0.000 n=10)
geomean                                     76.46µ        75.91µ       -0.72%
```